Delete everything that was replaced in the previous commit.
